feat: refatorada logica de despesa.

+++ Alugueis_API/Models/DespesaModel.cs
@@ -5,7 +5,7 @@
         public int CodDespesa { get; set; }
         public int CodTipoDespesa { get; set; }
         public TipoDespesa TipoDespesa { get; set; }
-        public double VrlTotalDespesa { get; set; }
+        public float VrlTotalDespesa { get; set; }
         public DateTime DataDespesa { get; set; }
         public DateTime CompetenciaMes { get; set; }
         public ICollection<DespesaRateio> Rateios { get; set; }
